@@ -1,26 +1,20 @@
 name: twig_dart
-version: 2.0.2
+version: 3.0.0
 description:  A simple server-side HTML templating engine for Dart. Comparable to Blade or Liquid.
 author: Isaque Neves <insinfo2008@gmail.com>
 homepage: https://github.com/insinfo/twig_dart
 environment:
-  sdk: ">=2.0.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   args: ^2.1.0 #Null safety
   charcode: ^1.2.0 #Null safety
   source_span: ^1.8.1 #Null safety
   string_scanner: ^1.1.0 #Null safety
-  matcher: ^0.12.10 #Null safety
-  #args: ^1.0.0
-  #charcode: ^1.0.0 
-  #source_span: ^1.0.0
-  #string_scanner: ^1.0.0
-
-  #symbol_table: ^2.0.0
-  #code_buffer: ^1.0.0
+  matcher: ^0.12.10 #Null safety 
   essential_code_buffer: ^2.0.1 #Null safety
   essential_symbol_table: ^2.0.1 #Null safety
+  collection: ^1.15.0-nullsafety.4
 
 dev_dependencies:
   pedantic: ^1.11.0 #Null safety
