@@ -6,22 +6,22 @@ import 'string.dart';
 import 'token.dart';
 
 class Attribute extends AstNode {
-  final Identifier id;
-  final StringLiteral string;
-  final Token equals, nequ;
-  final Expression value;
+  final Identifier? id;
+  final StringLiteral? string;
+  final Token? equals, nequ;
+  final Expression? value;
 
   Attribute(this.id, this.string, this.equals, this.nequ, this.value);
 
   bool get isRaw => nequ != null;
 
-  Expression get nameNode => id ?? string;
+  Expression? get nameNode => id ?? string;
 
-  String get name => string?.value ?? id.name;
+  String get name => string?.value ?? id!.name;
 
   @override
-  FileSpan get span {
-    if (equals == null) return nameNode.span;
-    return nameNode.span.expand(equals?.span ?? nequ.span).expand(value.span);
+  FileSpan? get span {
+    if (equals == null) return nameNode!.span;
+    return nameNode!.span!.expand(equals?.span ?? nequ!.span!).expand(value!.span!);
   }
 }
