@@ -7,19 +7,19 @@ import 'identifier.dart';
 import 'token.dart';
 
 class MemberExpression extends Expression {
-  final Expression expression;
-  final Token op;
+  final Expression? expression;
+  final Token? op;
   final Identifier name;
 
   MemberExpression(this.expression, this.op, this.name);
 
   @override
-  compute(SymbolTable scope) {
-    var target = expression.compute(scope);
-    if (op.span.text == '?.' && target == null) return null;
+  compute(SymbolTable? scope) {
+    var target = expression!.compute(scope);
+    if (op!.span!.text == '?.' && target == null) return null;
     return reflect(target).getField(Symbol(name.name)).reflectee;
   }
 
   @override
-  FileSpan get span => expression.span.expand(op.span).expand(name.span);
+  FileSpan get span => expression!.span!.expand(op!.span!).expand(name.span!);
 }
