@@ -7,17 +7,17 @@ import 'expression.dart';
 import 'token.dart';
 
 class StringLiteral extends Literal {
-  final Token string;
+  final Token? string;
   final String value;
 
   StringLiteral(this.string, this.value);
 
   static String parseValue(Token string) {
-    var text = string.span.text.substring(1, string.span.text.length - 1);
+    var text = string.span!.text.substring(1, string.span!.text.length - 1);
     var codeUnits = text.codeUnits;
     var buf = StringBuffer();
 
-    for (int i = 0; i < codeUnits.length; i++) {
+    for (var i = 0; i < codeUnits.length; i++) {
       var ch = codeUnits[i];
 
       if (ch == $backslash) {
@@ -63,10 +63,10 @@ class StringLiteral extends Literal {
   }
 
   @override
-  compute(SymbolTable scope) {
+  compute(SymbolTable? scope) {
     return value;
   }
 
   @override
-  FileSpan get span => string.span;
+  FileSpan? get span => string!.span;
 }
