@@ -3,7 +3,7 @@ import 'expression.dart';
 import 'token.dart';
 
 class Array extends Expression {
-  final Token lBracket, rBracket;
+  final Token? lBracket, rBracket;
   final List<Expression> items;
 
   Array(this.lBracket, this.rBracket, this.items);
@@ -13,24 +13,24 @@ class Array extends Expression {
 
   @override
   FileSpan get span {
-    return items.fold<FileSpan>(lBracket.span, (out, i) => out.expand(i.span)).expand(rBracket.span);
+    return items.fold<FileSpan?>(lBracket!.span, (out, i) => out!.expand(i.span!))!.expand(rBracket!.span!);
   }
 }
 
 class IndexerExpression extends Expression {
-  final Expression target, indexer;
-  final Token lBracket, rBracket;
+  final Expression? target, indexer;
+  final Token? lBracket, rBracket;
 
   IndexerExpression(this.target, this.lBracket, this.indexer, this.rBracket);
 
   @override
   FileSpan get span {
-    return target.span.expand(lBracket.span).expand(indexer.span).expand(rBracket.span);
+    return target!.span!.expand(lBracket!.span!).expand(indexer!.span!).expand(rBracket!.span!);
   }
 
   @override
   dynamic compute(scope) {
-    var a = target.compute(scope), b = indexer.compute(scope);
+    var a = target!.compute(scope), b = indexer!.compute(scope);
     return a[b];
   }
 }
