@@ -1,5 +1,5 @@
 import 'package:source_span/source_span.dart';
 
 abstract class AstNode {
-  FileSpan get span;
+  FileSpan? get span;
 }
