@@ -1,3 +1,4 @@
+import 'package:collection/collection.dart' show IterableExtension;
 import 'package:source_span/source_span.dart';
 import 'ast_node.dart';
 import 'attribute.dart';
@@ -12,7 +13,7 @@ class TextNode extends ElementChild {
   TextNode(this.text);
 
   @override
-  FileSpan get span => text.span;
+  FileSpan? get span => text.span;
 }
 
 abstract class Element extends ElementChild {
@@ -38,12 +39,12 @@ abstract class Element extends ElementChild {
 
   Iterable<ElementChild> get children;
 
-  Attribute getAttribute(String name) =>
-      attributes.firstWhere((a) => a.name == name, orElse: () => null);
+  Attribute? getAttribute(String name) =>
+      attributes.firstWhereOrNull((a) => a.name == name);
 }
 
 class SelfClosingElement extends Element {
-  final Token lt, slash, gt;
+  final Token? lt, slash, gt;
 
   final Identifier tagName;
 
@@ -58,15 +59,15 @@ class SelfClosingElement extends Element {
   @override
   FileSpan get span {
     var start = attributes.fold<FileSpan>(
-        lt.span.expand(tagName.span), (out, a) => out.expand(a.span));
+        lt!.span!.expand(tagName.span!), (out, a) => out.expand(a.span!));
     return slash != null
-        ? start.expand(slash.span).expand(gt.span)
-        : start.expand(gt.span);
+        ? start.expand(slash!.span!).expand(gt!.span!)
+        : start.expand(gt!.span!);
   }
 }
 
 class RegularElement extends Element {
-  final Token lt, gt, lt2, slash, gt2;
+  final Token? lt, gt, lt2, slash, gt2;
 
   final Identifier tagName, tagName2;
 
@@ -81,16 +82,16 @@ class RegularElement extends Element {
   FileSpan get span {
     var openingTag = attributes
         .fold<FileSpan>(
-            lt.span.expand(tagName.span), (out, a) => out.expand(a.span))
-        .expand(gt.span);
+            lt!.span!.expand(tagName.span!), (out, a) => out.expand(a.span!))
+        .expand(gt!.span!);
 
     if (gt2 == null) return openingTag;
 
     return children
-        .fold<FileSpan>(openingTag, (out, c) => out.expand(c.span))
-        .expand(lt2.span)
-        .expand(slash.span)
-        .expand(tagName2.span)
-        .expand(gt2.span);
+        .fold<FileSpan>(openingTag, (out, c) => out.expand(c.span!))
+        .expand(lt2!.span!)
+        .expand(slash!.span!)
+        .expand(tagName2.span!)
+        .expand(gt2!.span!);
   }
 }
