@@ -4,15 +4,15 @@ import 'expression.dart';
 import 'token.dart';
 
 class Interpolation extends ElementChild {
-  final Token doubleCurlyL, doubleCurlyR;
+  final Token? doubleCurlyL, doubleCurlyR;
   final Expression expression;
 
   Interpolation(this.doubleCurlyL, this.expression, this.doubleCurlyR);
 
-  bool get isRaw => doubleCurlyL.span.text.endsWith('-');
+  bool get isRaw => doubleCurlyL!.span!.text.endsWith('-');
 
   @override
   FileSpan get span {
-    return doubleCurlyL.span.expand(expression.span).expand(doubleCurlyR.span);
+    return doubleCurlyL!.span!.expand(expression.span!).expand(doubleCurlyR!.span!);
   }
 }
