@@ -1,4 +1,4 @@
-part of jael.src.text.parselet;
+part of twig.src.text.parselet;
 
 const Map<TokenType, PrefixParselet> prefixParselets = {
   TokenType.exclamation: NotParselet(),
@@ -16,12 +16,12 @@ class NotParselet implements PrefixParselet {
   const NotParselet();
 
   @override
-  Expression parse(Parser parser, Token token) {
+  Expression parse(Parser parser, Token? token) {
     var expression = parser.parseExpression(0);
 
     if (expression == null) {
       parser.errors.add(TwigDartError(
-          TwigDartErrorSeverity.error, 'Missing expression after "!" in negation expression.', token.span));
+          TwigDartErrorSeverity.error, 'Missing expression after "!" in negation expression.', token!.span));
     }
 
     return Negation(token, expression);
@@ -32,12 +32,12 @@ class NewParselet implements PrefixParselet {
   const NewParselet();
 
   @override
-  Expression parse(Parser parser, Token token) {
+  Expression? parse(Parser parser, Token? token) {
     var call = parser.parseExpression(0);
 
     if (call == null) {
       parser.errors.add(TwigDartError(
-          TwigDartErrorSeverity.error, '"new" must precede a call expression. Nothing was found.', call.span));
+          TwigDartErrorSeverity.error, '"new" must precede a call expression. Nothing was found.', call!.span));
       return null;
     } else if (call is Call) {
       return NewExpression(token, call);
@@ -53,30 +53,30 @@ class NumberParselet implements PrefixParselet {
   const NumberParselet();
 
   @override
-  Expression parse(Parser parser, Token token) => NumberLiteral(token);
+  Expression parse(Parser parser, Token? token) => NumberLiteral(token);
 }
 
 class HexParselet implements PrefixParselet {
   const HexParselet();
 
   @override
-  Expression parse(Parser parser, Token token) => HexLiteral(token);
+  Expression parse(Parser parser, Token? token) => HexLiteral(token);
 }
 
 class StringParselet implements PrefixParselet {
   const StringParselet();
 
   @override
-  Expression parse(Parser parser, Token token) => StringLiteral(token, StringLiteral.parseValue(token));
+  Expression parse(Parser parser, Token? token) => StringLiteral(token, StringLiteral.parseValue(token!));
 }
 
 class ArrayParselet implements PrefixParselet {
   const ArrayParselet();
 
   @override
-  Expression parse(Parser parser, Token token) {
+  Expression? parse(Parser parser, Token? token) {
     List<Expression> items = [];
-    Expression item = parser.parseExpression(0);
+    Expression? item = parser.parseExpression(0);
 
     while (item != null) {
       items.add(item);
@@ -87,7 +87,7 @@ class ArrayParselet implements PrefixParselet {
 
     if (!parser.next(TokenType.rBracket)) {
       var lastSpan = items.isEmpty ? null : items.last.span;
-      lastSpan ??= token.span;
+      lastSpan ??= token!.span;
       parser.errors
           .add(TwigDartError(TwigDartErrorSeverity.error, 'Missing "]" to terminate array literal.', lastSpan));
       return null;
@@ -101,7 +101,7 @@ class MapParselet implements PrefixParselet {
   const MapParselet();
 
   @override
-  Expression parse(Parser parser, Token token) {
+  Expression? parse(Parser parser, Token? token) {
     var pairs = <KeyValuePair>[];
     var pair = parser.parseKeyValuePair();
 
@@ -113,7 +113,7 @@ class MapParselet implements PrefixParselet {
     }
 
     if (!parser.next(TokenType.rCurly)) {
-      var lastSpan = pairs.isEmpty ? token.span : pairs.last.span;
+      var lastSpan = pairs.isEmpty ? token!.span : pairs.last.span;
       parser.errors.add(TwigDartError(TwigDartErrorSeverity.error, 'Missing "}" in map literal.', lastSpan));
       return null;
     }
@@ -126,18 +126,18 @@ class IdentifierParselet implements PrefixParselet {
   const IdentifierParselet();
 
   @override
-  Expression parse(Parser parser, Token token) => Identifier(token);
+  Expression parse(Parser parser, Token? token) => Identifier(token);
 }
 
 class ParenthesisParselet implements PrefixParselet {
   const ParenthesisParselet();
 
   @override
-  Expression parse(Parser parser, Token token) {
+  Expression? parse(Parser parser, Token? token) {
     var expression = parser.parseExpression(0);
 
     if (expression == null) {
-      parser.errors.add(TwigDartError(TwigDartErrorSeverity.error, 'Missing expression after "(".', token.span));
+      parser.errors.add(TwigDartError(TwigDartErrorSeverity.error, 'Missing expression after "(".', token!.span));
       return null;
     }
 
