@@ -3,18 +3,18 @@ import 'package:twig_dart/twig_dart.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('attributes', () {
+  test('attributes 2', () {
     var doc = parseDSX('''
     <foo bar="baz" yes={no} />
-    ''');
+    ''')!;
 
     var foo = doc.root as SelfClosingElement;
     expect(foo.tagName.name, 'foo');
     expect(foo.attributes, hasLength(2));
     expect(foo.getAttribute('bar'), isNotNull);
     expect(foo.getAttribute('yes'), isNotNull);
-    expect(foo.getAttribute('bar').value.compute(null), 'baz');
-    expect(foo.getAttribute('yes').value.compute(SymbolTable(values: {'no': 'maybe'})), 'maybe');
+    expect(foo.getAttribute('bar')!.value!.compute(null), 'baz');
+    expect(foo.getAttribute('yes')!.value!.compute(SymbolTable(values: {'no': 'maybe'})), 'maybe');
   });
 
   test('children', () {
@@ -22,7 +22,7 @@ void main() {
     <foo bar="baz" yes={no}>
       <bar>{24 * 3}</bar>
     </foo>
-    ''');
+    ''')!;
 
     var bar = doc.root.children.first as RegularElement;
     expect(bar.tagName.name, 'bar');
@@ -32,6 +32,6 @@ void main() {
   });
 }
 
-Document parseDSX(String text) {
+Document? parseDSX(String text) {
   return parseDocument(text, sourceUrl: 'test.dsx', asDSX: true, onError: (e) => throw e);
 }
